# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NeuroWealth API
  version: 1.0.0
  description: |
    REST API for the NeuroWealth platform — AI-assisted portfolio management
    backed by Stellar smart contracts.

    **Versioning:** All endpoints are served under an explicit version prefix
    (`/api/v1/*`). The legacy unversioned paths (`/api/*`) remain available as
    deprecated aliases and return `Deprecation`/`Sunset` headers. Every response
    includes an `X-API-Version` header. See `docs/api-versioning.md`.

    **Breaking-change policy:** This API follows semantic versioning. Breaking
    changes (removed fields, changed response shapes, new required parameters)
    increment the major version and are announced at least two weeks before release.
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: '3000'
  - url: https://api.neurowealth.app
    description: Production
...
security: []
paths:
...
  /api/v1/deposit:
    post:
      tags: [deposit]
      operationId: initiateDeposit
      summary: Initiate an on-chain deposit
      description: |
        Submits a deposit transaction to the active protocol on behalf of the
        authenticated user. The user must match the userId in the request body,
        OR the caller must hold an active sub-account with DEPOSIT permission
        for the target user.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, amount, assetSymbol]
              properties:
                userId:
                  type: string
                  format: uuid
                  description: User ID (UUID v4)
                amount:
                  type: number
                  minimum: 0
                  exclusiveMinimum: true
                  description: Deposit amount
                assetSymbol:
                  type: string
                  minLength: 1
                  description: Asset symbol (e.g. USDC, XLM)
                protocolName:
                  type: string
                  description: Target protocol (optional, uses active protocol if omitted)
                memo:
                  type: string
                  maxLength: 280
                  description: Optional memo
            example:
              userId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
              amount: 1000.00
              assetSymbol: USDC
              protocolName: Aave
      responses:
        '201':
          description: Deposit initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
              example:
                txHash: a1b2c3d4e5f6...
                status: CONFIRMED
                transaction:
                  id: tx-001
                  txHash: a1b2c3d4e5f6...
                  status: CONFIRMED
                  amount: 1000.00
                  assetSymbol: USDC
                  protocolName: Aave
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Duplicate transaction hash
          $ref: '#/components/responses/Conflict'

  # ── Recurring Deposit ─────────────────────────────────────────────────────
  /api/v1/deposit/recurring:
    post:
      tags: [deposit]
      operationId: createRecurringDeposit
      summary: Create a recurring deposit plan
      description: |
        Creates a scheduled recurring deposit. Requires explicit confirmation.
        The deposit will execute automatically on the specified cadence.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, amount, assetSymbol, cadence, confirmed]
              properties:
                userId:
                  type: string
                  format: uuid
                amount:
                  type: number
                  exclusiveMinimum: true
                assetSymbol:
                  type: string
                  minLength: 1
                cadence:
                  type: string
                  enum: [WEEKLY, BIWEEKLY, MONTHLY]
                confirmed:
                  type: boolean
                  enum: [true]
                  description: Must be true to confirm the recurring deposit
            example:
              userId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
              amount: 50
              assetSymbol: USDC
              cadence: WEEKLY
              confirmed: true
      responses:
        '201':
          description: Recurring deposit plan created
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    $ref: '#/components/schemas/RecurringDepositPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/deposit/recurring/by-user/{userId}:
    get:
      tags: [deposit]
      operationId: listRecurringDeposits
      summary: List recurring deposit plans for a user
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of recurring deposit plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecurringDepositPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/deposit/recurring/{id}:
    patch:
      tags: [deposit]
      operationId: updateRecurringDeposit
      summary: Update a recurring deposit plan
      description: Pause, resume, or update the amount/cadence of a plan.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  exclusiveMinimum: true
                cadence:
                  type: string
                  enum: [WEEKLY, BIWEEKLY, MONTHLY]
                status:
                  type: string
                  enum: [ACTIVE, PAUSED, CANCELLED]
      responses:
        '200':
          description: Plan updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    $ref: '#/components/schemas/RecurringDepositPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [deposit]
      operationId: cancelRecurringDeposit
      summary: Cancel a recurring deposit plan
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    $ref: '#/components/schemas/RecurringDepositPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Withdraw ───────────────────────────────────────────────────────────────
...
  /api/v1/alerts:
    post:
      tags: [alerts]
      operationId: createAlertRule
      summary: Create an alert rule
      description: >
        Creates a user-defined price/yield alert rule owned by the authenticated
        caller. The rule is evaluated on a schedule; when its comparator
        condition holds against the live metric and it is outside its cooldown
        window, a notification is delivered over the chosen channel(s). A
        `PROTOCOL_APY` rule must name a `protocolName`; the other metrics must
        not. See `docs/ALERTS.md` for metric semantics and the drawdown window.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertRuleRequest'
      responses:
        '201':
          description: The created alert rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # GET is keyed by the owning user id; PATCH/DELETE by the rule id. They share
  # one URL template (disambiguated only by HTTP method in the Express router),
  # so OpenAPI represents them under a single path item with a generic {id}
  # whose meaning is documented per-operation.
  /api/v1/alerts/{id}:
    get:
      tags: [alerts]
      operationId: listAlertRules
      summary: List a user's alert rules
      description: >
        Lists the alert rules owned by the user, newest first. Owner-scoped — a
        caller may only read their own rules. Here the path parameter is the
        owning **user id** (must match the authenticated caller).
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The owning user's id (must match the authenticated caller)
      responses:
        '200':
          description: Alert rules for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRuleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [alerts]
      operationId: updateAlertRule
      summary: Update an alert rule
      description: >
        Updates a rule the caller owns. Here the path parameter is the **rule
        id**. The `PROTOCOL_APY` ↔ `protocolName` pairing is enforced against
        the merged (stored + patch) state.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Alert rule id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlertRuleRequest'
      responses:
        '200':
          description: The updated alert rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [alerts]
      operationId: deleteAlertRule
      summary: Delete an alert rule
      description: Deletes a rule the caller owns. Future evaluations stop immediately.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Alert rule id
      responses:
        '204':
          description: Deleted (no content)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Sub-Accounts ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued by `POST /api/v1/auth/verify`.
        Include as `Authorization: Bearer <token>`.
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: API-Key
      description: >
        Admin API key. Can be provided either as
        `Authorization: Bearer <token>` or as the legacy
        `X-Admin-Token` header. Issued via `POST /api/v1/admin/keys`.
    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: >
        Service-to-service token for internal endpoints (e.g., agent status,
        Prometheus metrics). Alternative auth methods include IP allowlisting
        or admin Bearer token.
  schemas:
...
    TransactionResponse:
      type: object
      properties:
        txHash:
          type: string
        status:
          type: string
        transaction:
          type: object
          properties:
            id:
              type: string
            txHash:
              type: string
            status:
              type: string
            amount:
              type: number
            assetSymbol:
              type: string
            protocolName:
              type: string
              nullable: true
            actingAsUserId:
              type: string
              format: uuid
              nullable: true
              description: >
                Non-null when the action was performed by a parent on behalf
                of a child via a sub-account delegation.

    # ── Recurring Deposit ──────────────────────────────────────────────────
    RecurringDepositPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: string
          description: Decimal amount
        assetSymbol:
          type: string
        cadence:
          type: string
          enum: [WEEKLY, BIWEEKLY, MONTHLY]
        nextRunAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [ACTIVE, PAUSED, CANCELLED]
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        lastRunStatus:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Protocol ──────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Unauthorized
        details:
          description: Optional structured detail (array or object)
          oneOf:
            - type: array
              items: {}
            - type: object

    # ── Sub-Accounts ─────────────────────────────────────────────────────────
...
    AlertMetric:
      type: string
      enum: [PROTOCOL_APY, PORTFOLIO_VALUE, POSITION_DRAWDOWN]
      description: >
        What the rule watches. PROTOCOL_APY = a named protocol's supply APY in
        percent; PORTFOLIO_VALUE = the user's total active-position value in USD;
        POSITION_DRAWDOWN = percentage decline of that portfolio value from its
        rolling 30-day peak (see docs/ALERTS.md).
    AlertComparator:
      type: string
      enum: [LT, LTE, GT, GTE]
      description: Comparison of the observed value against the threshold.
    AlertDeliveryChannel:
      type: string
      enum: [WEBHOOK, WHATSAPP, BOTH]
      description: Where a triggered alert is delivered.
...
    AlertRule:
      type: object
      required:
        - id
        - userId
        - metric
        - comparator
        - threshold
        - deliveryChannel
        - cooldownMinutes
        - isActive
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        metric:
          $ref: '#/components/schemas/AlertMetric'
        protocolName:
          type: string
          nullable: true
          description: Required when metric is PROTOCOL_APY; null otherwise.
          example: Blend
        comparator:
          $ref: '#/components/schemas/AlertComparator'
        threshold:
          type: number
          description: >
            Percent for PROTOCOL_APY/POSITION_DRAWDOWN, USD for PORTFOLIO_VALUE.
          example: 5
        deliveryChannel:
          $ref: '#/components/schemas/AlertDeliveryChannel'
        cooldownMinutes:
          type: integer
          minimum: 1
          maximum: 10080
          default: 60
          description: Minimum minutes between repeat notifications while the condition holds.
        lastFiredAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
          description: >
            A PROTOCOL_APY rule is auto-set to false if its protocol is delisted
            (no rate data), rather than evaluating against stale data.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAlertRuleRequest:
      type: object
      required: [metric, comparator, threshold, deliveryChannel]
      properties:
        metric:
          $ref: '#/components/schemas/AlertMetric'
        protocolName:
          type: string
          description: Required when metric is PROTOCOL_APY; rejected otherwise.
          example: Blend
        comparator:
          $ref: '#/components/schemas/AlertComparator'
        threshold:
          type: number
          example: 5
        deliveryChannel:
          $ref: '#/components/schemas/AlertDeliveryChannel'
        cooldownMinutes:
          type: integer
          minimum: 1
          maximum: 10080
          default: 60
    UpdateAlertRuleRequest:
      type: object
      minProperties: 1
      description: Partial update — at least one field must be provided.
      properties:
        metric:
          $ref: '#/components/schemas/AlertMetric'
        protocolName:
          type: string
          nullable: true
        comparator:
          $ref: '#/components/schemas/AlertComparator'
        threshold:
          type: number
        deliveryChannel:
          $ref: '#/components/schemas/AlertDeliveryChannel'
        cooldownMinutes:
          type: integer
          minimum: 1
          maximum: 10080
        isActive:
          type: boolean
    AlertRuleListResponse:
      type: object
      required: [rules]
      properties:
        rules:
          type: array
          items:
            $ref: '#/components/schemas/AlertRule'
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
...
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Validation error
            details:
              - field: amount
                message: Must be a positive number
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not found
    Conflict:
      description: Resource conflict (e.g. duplicate)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Conflict